+++ sessionize-base/src/main/kotlin/info/metadude/kotlin/library/sessionize/SessionizeApi.kt
@@ -1,0 +1,12 @@
+package info.metadude.kotlin.library.sessionize
+
+import okhttp3.Call
+
+interface SessionizeApi {
+
+    fun provideSessionizeService(
+        baseUrl: String,
+        callFactory: Call.Factory,
+    ): SessionizeService
+
+}